Ignore incomplete tail commands and allow for null padding
This commit allows for BuntDB to load data files that were
previously considered invalid or corrupted.

Now when the data file ends with an incomplete command, the data
will be truncated at the end of the previously success command.
Also when a null control character is encountered instead of an
asterix, which indicates the start of a command, the null is
ignored and the cursor moves to the next byte. This allows for
null padding at the head and the tail.

Fixes #71
https://github.com/tidwall/tile38/issues/600